requirements.txt
dataclasses==0.8
importlib-metadata==4.8.3
pyserial==3.5
PyVISA==1.11.3
PyVISA-py==0.5.2
typing-extensions==4.1.1
zipp==3.6.0
